+++ src/main/java/PP/alquilerVehiculo/servicio/BaseService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface BaseService<E> {
     public List<E> findAll() throws Exception;
 
-    public E findById(long id) throws Exception;
+    public E findById(Integer id) throws Exception;
 
     public E save(E entity) throws Exception;
 
-    public E update(long id, E entity) throws Exception;
+    public E update(Integer id, E entity) throws Exception;
 
-    public void deleteById(long id) throws Exception;
+    public void deleteById(Integer id) throws Exception;
 }
